Moved utility methods into utils.py
